Update google cloud library
