+++ libraries/google_cloud/tuples.py
