+++ requirements.txt
@@ -7,8 +7,5 @@ django-redis==5.4.0
 psycopg2-binary==2.9.10
 
 Pillow==11.1.0
-# S3 storage
-paramiko==3.5.0
-boto3==1.36.7
 # Google drive storage
-# google-api-python-client==2.117.0
+google-api-python-client==2.117.0
